[PL-9865]: Migrate instance sync for PCF deployments to perpetual tasks. (#12265)
* Added Comments

* Added Instance sync as a perpetual task

* Linked task creation to version info api

* Working code on the manager side. The delegate fn is getting called at the desired interval.

* Milestone #1: Delegate is reporting to manager the New instance count

* Refactored the code for PCFInstance Handler

* reverted version info file

* Everything working aftr updating the instance Sync

* Reverting 19-grpc changes

* Revert "Reverting 19-grpc changes"

This reverts commit 67c10bf80b7921012d94e5dce49cb442546e9cc2.

* Reverting 19-grpc changes

* fClang formatting

* Added check for Failure in case of Delegate Agent Response.

* Deleted the file

* Moved constants to one class.

* Minor refactoring

* Changed account ID in the executor

* Revert "Reverting 19-grpc changes"

This reverts commit b7008ecf61a9d23cb1552671d08e7bceeb588fbd.

* Temp commit

* Revert files

* Revert "Temp commit"

This reverts commit 7f7cd17dac274684c2f36aa198fe1a4dcee71f2b.

* Addressed code review commments: Added Singleton annotation and few minor changes

* Added feature flag

* Added Control from feature flag for adding the perpetual tasks.

* Added another FF for skipping the Instance sync via iterator

* Refactored the PcfCommandResponse to be able to serialize the subtypes as well.

* Fixed NPE error in the Manual instance sync flow

* Added error handling in case of pcfInstanceSyncResponse and HarnessUserGroup Check in the API

* Chnaged the frequency of perpetual taks back to 10 mins

* Refactored code and extracted out the ff control to the classes

* Minor refactoring: self CR

* Added changes for task reset

* Sending error code as PT response.

* Addressed final CR, Vikas

* Fixed checkstyle issues

* Fixed checks

* Fixed test and clang formatting

* Fixed checks issues and the mongo index

* Fixed changes for CRLF checks

* Added test for PcfInstanceSyncPerpTaskClient

* Added test for PCFInstanceSyncPerpetualTaskControllerTest

* Added test for PCFInstanceSyncPerpetualTaskControllerTest

* Ran clang format for 19-grpc

* Clang formatting

* Clang-formatting

+++ 71-rest/src/main/java/software/wings/service/impl/instance/PcfInstanceHandler.java
@@ -13,7 +13,11 @@ import com.google.inject.Singleton;
 
 import io.harness.data.structure.EmptyPredicate;
 import io.harness.exception.WingsException;
+import io.harness.perpetualtask.PerpetualTaskClientContext;
+import io.harness.perpetualtask.PerpetualTaskService;
+import io.harness.perpetualtask.internal.PerpetualTaskRecord;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
 import software.wings.api.DeploymentInfo;
 import software.wings.api.DeploymentSummary;
 import software.wings.api.PcfDeploymentInfo;
@@ -34,6 +38,11 @@ import software.wings.beans.infrastructure.instance.key.PcfInstanceKey;
 import software.wings.beans.infrastructure.instance.key.deployment.DeploymentKey;
 import software.wings.beans.infrastructure.instance.key.deployment.PcfDeploymentKey;
 import software.wings.helpers.ext.pcf.PcfAppNotFoundException;
+import software.wings.helpers.ext.pcf.response.PcfCommandExecutionResponse;
+import software.wings.helpers.ext.pcf.response.PcfInstanceSyncResponse;
+import software.wings.service.InstanceSyncController;
+import software.wings.service.InstanceSyncController.InstanceSyncFlow;
+import software.wings.service.PcfInstanceSyncConstants;
 import software.wings.service.impl.PcfHelperService;
 import software.wings.sm.PhaseStepExecutionSummary;
 import software.wings.sm.StepExecutionSummary;
@@ -53,24 +62,64 @@ import java.util.stream.Collectors;
 @Slf4j
 public class PcfInstanceHandler extends InstanceHandler {
   @Inject private PcfHelperService pcfHelperService;
+  @Inject private PerpetualTaskService perpetualTaskService;
+  @Inject private InstanceSyncController instanceSyncController;
 
   @Override
   public void syncInstances(String appId, String infraMappingId) {
     Multimap<String, Instance> pcfAppNameInstanceMap = ArrayListMultimap.create();
     syncInstancesInternal(appId, infraMappingId, pcfAppNameInstanceMap, null, false,
-        OnDemandRollbackInfo.builder().onDemandRollback(false).build());
+        OnDemandRollbackInfo.builder().onDemandRollback(false).build(), null);
   }
 
+  public void processInstanceSyncResponse(
+      PcfCommandExecutionResponse pcfCommandExecutionResponse, String perpetualTaskId) throws PcfAppNotFoundException {
+    PcfInstanceSyncResponse pcfInstanceSyncResponse =
+        (PcfInstanceSyncResponse) pcfCommandExecutionResponse.getPcfCommandResponse();
+
+    logger.info("Received instance sync response {}, perpetual taskId: {}",
+        CollectionUtils.size(pcfInstanceSyncResponse.getInstanceIndices()), perpetualTaskId);
+    PerpetualTaskRecord taskContext = perpetualTaskService.getTaskRecord(perpetualTaskId);
+    PerpetualTaskClientContext clientParamMap = taskContext.getClientContext();
+
+    Map<String, String> clientParams = clientParamMap.getClientParams();
+
+    String infraId = clientParams.get(InfrastructureMapping.ID_KEY);
+    String applicationId = clientParams.get(PcfInstanceSyncConstants.APPLICATION_ID);
+
+    Multimap<String, Instance> pcfAppNameInstanceMap = ArrayListMultimap.create();
+    try {
+      syncInstancesInternal(applicationId, infraId, pcfAppNameInstanceMap, null, false,
+          OnDemandRollbackInfo.builder().onDemandRollback(false).build(), pcfCommandExecutionResponse);
+    } catch (Exception ex) {
+      logger.warn("Instance sync failed for infraMapping", ex);
+      String errorMsg;
+      if (ex instanceof WingsException) {
+        errorMsg = ex.getMessage();
+      } else {
+        errorMsg = ex.getMessage() != null ? ex.getMessage() : "Unknown error";
+      }
+      InfrastructureMapping infraMapping = infraMappingService.get(applicationId, infraId);
+      instanceService.handleSyncFailure(applicationId, infraMapping.getServiceId(), infraMapping.getEnvId(), infraId,
+          infraMapping.getDisplayName(), System.currentTimeMillis(), errorMsg);
+    }
+  }
   /**
-   *
-   * @param appId
+   *  @param appId
    * @param infraMappingId
    * @param pcfAppNameInstanceMap  key - pcfAppName     value - Instances
+   * @param pcfCommandExecutionResponse
    */
   private void syncInstancesInternal(String appId, String infraMappingId,
       Multimap<String, Instance> pcfAppNameInstanceMap, List<DeploymentSummary> newDeploymentSummaries,
-      boolean rollback, OnDemandRollbackInfo onDemandRollbackInfo) {
-    logger.info("# Performing PCF Instance sync");
+      boolean rollback, OnDemandRollbackInfo onDemandRollbackInfo,
+      final PcfCommandExecutionResponse pcfCommandExecutionResponse) {
+    PcfInstanceSyncResponse pcfInstanceSyncResponse = pcfCommandExecutionResponse == null
+        ? null
+        : (PcfInstanceSyncResponse) pcfCommandExecutionResponse.getPcfCommandResponse();
+    InstanceSyncFlow instanceSyncFlow = getInstanceSyncFlow(pcfCommandExecutionResponse, newDeploymentSummaries);
+
+    logger.info("Performing PCF Instance sync. Caller is: {}, inframapping id: {}", instanceSyncFlow, infraMappingId);
     InfrastructureMapping infrastructureMapping = infraMappingService.get(appId, infraMappingId);
     notNullCheck("Infra mapping is null for id:" + infraMappingId, infrastructureMapping);
 
@@ -83,7 +132,8 @@ public class PcfInstanceHandler extends InstanceHandler {
 
     Map<String, DeploymentSummary> pcfAppNamesNewDeploymentSummaryMap = getDeploymentSummaryMap(newDeploymentSummaries);
 
-    loadPcfAppNameInstanceMap(appId, infraMappingId, pcfAppNameInstanceMap);
+    String applicationNameIfSupplied = (pcfInstanceSyncResponse == null) ? null : pcfInstanceSyncResponse.getName();
+    loadPcfAppNameInstanceMap(appId, infraMappingId, pcfAppNameInstanceMap, applicationNameIfSupplied);
 
     PcfInfrastructureMapping pcfInfrastructureMapping = (PcfInfrastructureMapping) infrastructureMapping;
     SettingAttribute settingAttribute = settingsService.get(pcfInfrastructureMapping.getComputeProviderSettingId());
@@ -98,8 +148,9 @@ public class PcfInstanceHandler extends InstanceHandler {
         boolean failedToRetrieveData = false;
         List<PcfInstanceInfo> latestpcfInstanceInfoList = new ArrayList<>();
         try {
-          latestpcfInstanceInfoList = pcfHelperService.getApplicationDetails(pcfApplicationName,
-              pcfInfrastructureMapping.getOrganization(), pcfInfrastructureMapping.getSpace(), pcfConfig);
+          latestpcfInstanceInfoList =
+              pcfHelperService.getApplicationDetails(pcfApplicationName, pcfInfrastructureMapping.getOrganization(),
+                  pcfInfrastructureMapping.getSpace(), pcfConfig, pcfCommandExecutionResponse);
         } catch (Exception e) {
           logger.warn("Error while fetching application details for PCFApplication", e);
 
@@ -152,11 +203,17 @@ public class PcfInstanceHandler extends InstanceHandler {
           });
           logger.info("Instances to be deleted {}", instanceIdsToBeDeleted.size());
 
-          logger.info("Total no of instances found in DB for AppId: {}, "
+          logger.info("Total no of instances found in DB wih caller: {} AppId: {}, "
                   + "No of instances in DB: {}, No of Running instances: {}, "
                   + "No of instances to be Added: {}, No of instances to be deleted: {}",
-              appId, instancesInDB.size(), latestPcfInstanceInfoMap.keySet().size(), instancesToBeAdded.size(),
-              instanceIdsToBeDeleted.size());
+              instanceSyncFlow, appId, instancesInDB.size(), latestPcfInstanceInfoMap.keySet().size(),
+              instancesToBeAdded.size(), instanceIdsToBeDeleted.size());
+
+          // Control if the DB will be updated by PERP_TASK or Iterator_INSTANCE_SYNC
+          if (!instanceSyncController.canUpdateDb(instanceSyncFlow, infrastructureMapping, getClass())) {
+            return;
+          }
+
           if (isNotEmpty(instanceIdsToBeDeleted)) {
             instanceService.delete(instanceIdsToBeDeleted);
           }
@@ -192,6 +249,25 @@ public class PcfInstanceHandler extends InstanceHandler {
           }
         }
       });
+    }
+  }
+
+  /**
+   * Returns the parent caller function.
+   * The function can be called from 3 different sources and the return will be:
+   *    In case of new deployment - NEW_DEPLOYMENT
+   *    From iterator sync from the manager side - ITERATOR_INSTANCE_SYNC
+   *    Called after the PT gets the instances in the delegate and calls the manager to process the response -
+   * PERPETUAL_TASK
+   */
+  private InstanceSyncFlow getInstanceSyncFlow(
+      PcfCommandExecutionResponse pcfInstanceSyncDelegateResponse, List<DeploymentSummary> newDeploymentSummaries) {
+    if (newDeploymentSummaries != null) {
+      return InstanceSyncFlow.NEW_DEPLOYMENT;
+    } else if (pcfInstanceSyncDelegateResponse != null) {
+      return InstanceSyncFlow.PERPETUAL_TASK;
+    } else {
+      return InstanceSyncFlow.ITERATOR_INSTANCE_SYNC;
     }
   }
 
@@ -237,14 +313,18 @@ public class PcfInstanceHandler extends InstanceHandler {
     return builder.build();
   }
 
-  private void loadPcfAppNameInstanceMap(
-      String appId, String infraMappingId, Multimap<String, Instance> pcfApplicationNameInstanceMap) {
+  private void loadPcfAppNameInstanceMap(String appId, String infraMappingId,
+      Multimap<String, Instance> pcfApplicationNameInstanceMap, String applicationNameIfSupplied) {
     List<Instance> instanceListInDBForInfraMapping = getInstances(appId, infraMappingId);
     for (Instance instance : instanceListInDBForInfraMapping) {
       InstanceInfo instanceInfo = instance.getInstanceInfo();
       if (instanceInfo instanceof PcfInstanceInfo) {
         PcfInstanceInfo pcfInstanceInfo = (PcfInstanceInfo) instanceInfo;
         String pcfAppName = pcfInstanceInfo.getPcfApplicationName();
+        // todo(aman) this logic needs to be reviewed.
+        if (isNotEmpty(applicationNameIfSupplied) && !pcfAppName.equals(applicationNameIfSupplied)) {
+          continue;
+        }
         pcfApplicationNameInstanceMap.put(pcfAppName, instance);
       } else {
         throw WingsException.builder()
@@ -265,7 +345,7 @@ public class PcfInstanceHandler extends InstanceHandler {
 
     syncInstancesInternal(deploymentSummaries.iterator().next().getAppId(),
         deploymentSummaries.iterator().next().getInfraMappingId(), pcfApplicationNameInstanceMap, deploymentSummaries,
-        rollback, onDemandRollbackInfo);
+        rollback, onDemandRollbackInfo, null);
   }
 
   @Override
